ws: Add (yet unsafe) client certificate authentication
Add a new pam_cockpit_cert PAM module for TLS client certificate based
authentication. This uses sssd's API to map a TLS certificate to a
user [1]. Use the sd-bus API for making the D-Bus calls, instead of
gdbus (glib/gio has too many side effects, which should be avoided in a
setuid root program) or libdbus (which would be a new dependency).
sd-bus is not yet available in RHEL/CentOS 7, so build it conditionally.

Introduce a new `tls-cert` authorization type into cockpit-ws that gets
enabled with the new `UnsafeClientCertAuthentication` option in
cockpit.conf. If this is enabled and the client sends a TLS client
certificate, call the new `cockpit-cert-session` session command instead
of `cockpit-session`. That will pass the certificate instead of the (yet
unknown) user name to the PAM stack, for pam_cockpit_cert.

This is currently unsafe: It trusts cockpit-ws to do the TLS
termination, in particular validating that the remote end really has the
corresponding private key. Other authentication methods
(cockpit-{session,ssh} don't have to trust ws and do the validation
themselves. But this is tricky to do with a TLS connection as it's not
possible to get back to the browser and ask for the certificate again.
This needs to be re-engineered more fundamentally. Until then this is a
proof of concept and remains undocumented in the man pages and guide.
The "Unsafe.." option name hints to that. Only document this in
authentication.md, which is just for developers and not shipped.

On Fedora/RHEL, ships this in a (possibly temporary) new package
cockpit-ws-cert-session, which only gets built for CI (`wip`), not for
releases. Don't ship it yet on Debian/Ubuntu.

Unfortunately the Chrome Devtools Protocol does not currently offer
client certificate import and selecting one for a page that requests a
certificate. Test this with curl instead, and describe in the comments
how to test this interactively.

[1] https://www.freeipa.org/page/V4/User_Certificates

Related: https://bugzilla.redhat.com/show_bug.cgi?id=1678465
Jira: COCKPIT-368
Closes #11421