test: Let image-prepare do a development build only upon request
Only enable image-prepare's automagic "`NODE_ENV=development` for
default OS" behaviour upon explicit request
(`--devel-build-on-default-os`), and set it for our verify scenarios in
test/run.

This retains the current behaviour in CI without hardcoding any OS name,
but calling test/image-prepare in a clean tree locally will then do a
standard (production) build. This is less surprising for a developer, as
they usually explicitly configure with `--enable-debug` or set
`$NODE_ENV`.

This will help us for downloading the pre-built dist tarballs from
GitHub. We want to get this working for the default OS (as that's the
one we most often review on), and not skip the download because
image-prepare wants to do a development build.