test: Extract dist/ and node_modules/ from downloaded dist tarballs
This is practical to immediately be able to run integration tests
without requiring `npm install`, and it should also help with not having
to rebuild the webpacks with `make` if there is no code change.

It is infuriatingly impossible to tell the tar CLI or Python's tarfile to
extract only 'cockpit-<version>/dist', so unpack the whole thing into a
temporary directory and move out the dirs that we want.