workflows: Move PR tests and statistics from webhook
Trigger integration tests after the build-dist workflow, so that they
can download the dist tarball instead of having to build it again.  This
also replaces the

  PR webhook → webhook AMQP queue → bot → test-scan → test AMQP queue

roundtrip with a more direct

  PR workflow → test AMQP queue

We also used the pull_request webhook to update the test statistics for
merged PRs. Replace this with the pr-statistics.yml workflow.

We have to do both of these in lockstep, as one cannot configure
webhooks to only fire for particular event types (in particular, only on
"closed").