test: Use real mouse events with Chromium in the shell
BiDi mouse events work fine on the top-level document (i.e. not in
frames). So let's use them to test the shell.

Generalize the "move mouse position to a harmless place" hack from
commit b5b8c6df31e26, so that we make the pixel snapshots without
focusing a random element on the page.

We can also run `TestPages.testPageStatus` again (reverts commit
529e3347aa11b6).

This just needs one little tweak in TestHostSwitching.testBasic's pixel
test. Our current reference does not expect any dialog item to be
highlighted (when the mouse is over it), so move the mouse out of the
dialog.