test: Wait for animation in TestHostSwitching
Opening the host switcher and the host edit view animates (expands) the
hosts menu for about half a second. When doing real clicks instead of
MouseEvent injection, this causes the target elements to move around
underneath the test, resulting in mis-clicks.

Avoid that by waiting for the animation to be over. Unfortunately
`ph_count_animations()` doesn't quite catch it completely in all cases,
so add an extra sleep for robustness. This is similar to what we do in
the pixel tests.